[Support] Move OptimizationLevel to the Support directory
Sometimes code depends upon the constant variables declared
in OptimizationLevel (e.g. see #156802), which then requires
them to add a dependency on the Passes directory. Given that
code is much more likely to depend upon Support than Passes
I believe it makes sense to move OptimizationLevel to
Support. The Passes component already depends upon Support.
