+++ mlir/lib/ExecutionEngine/OptUtils.cpp
@@ -15,10 +15,10 @@
 
 #include "llvm/Analysis/TargetTransformInfo.h"
 #include "llvm/IR/Module.h"
-#include "llvm/Passes/OptimizationLevel.h"
 #include "llvm/Passes/PassBuilder.h"
 #include "llvm/Support/Error.h"
 #include "llvm/Support/FormatVariadic.h"
+#include "llvm/Support/OptimizationLevel.h"
 #include "llvm/Target/TargetMachine.h"
 #include <optional>
